# api/docs/swagger.yaml
basePath: /api
definitions:
...
  middlewares.PermissionDenied:
    properties:
      error:
        type: string
    type: object
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
...
  models.System:
    properties:
      _:
        type: integer
      google_captcha_secret:
        type: string
      google_captcha_site_key:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      is_admin:
        type: boolean
      last_login:
        type: string
      uid:
        type: string
      updated_at:
        type: string
      username:
        type: string
    required:
    - is_admin
    - last_login
    - uid
    - username
    type: object
  models.UsersLookup:
    properties:
      uid:
        type: string
      username:
        type: string
    required:
    - uid
    - username
    type: object
...
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
  request.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_records:
        type: integer
    required:
    - page
    - size
    - total_records
    type: object
  system.ChangeUserPassword:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  system.ChangeUserPasswordBySelf:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  system.CreateUserData:
    properties:
      admin:
        type: boolean
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  system.GetSystemInitResponse:
    properties:
      google_captcha_site_key:
        type: string
    type: object
  system.GetSystemResponse:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
    type: object
  system.LoginData:
    properties:
      password:
        example: doe123456
        maxLength: 16
        minLength: 2
        type: string
      remember_me:
        type: boolean
      token:
        type: string
      username:
        example: john_doe
        maxLength: 16
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  system.PatchSystemUpdateData:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
    required:
    - google_captcha_secret_key
    - google_captcha_site_key
    type: object
  system.SetupSystem:
    properties:
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
      password:
        maxLength: 16
        minLength: 4
        type: string
      username:
        maxLength: 16
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  system.SetupSystemResponse:
    properties:
      system:
        $ref: '#/definitions/models.System'
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    required:
    - system
    - token
    - user
    type: object
  system.UserLoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    required:
    - token
    - user
    type: object
  system.UsersResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - meta
    type: object
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /system:
    get:
      consumes:
      - application/json
      description: Get panel System Config
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.GetSystemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Get panel System Config
      tags:
      - System
    patch:
      consumes:
      - application/json
      description: Update panel System Config
      parameters:
      - description: update system config data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/system.PatchSystemUpdateData'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.GetSystemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Update panel System Config
      tags:
      - System
  /system/init:
    get:
      consumes:
      - application/json
      description: Get panel System init Config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.GetSystemInitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Get panel System init Config
      tags:
      - System
  /system/setup:
    post:
      consumes:
      - application/json
      description: Setup user and system config
      parameters:
      - description: system setup data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/system.SetupSystem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/system.SetupSystemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Setup user and system config
      tags:
      - System
  /system/users:
    get:
      consumes:
      - application/json
      description: List of Admin or simple users
      parameters:
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.UsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: List of Admin or simple users
      tags:
      - System(Users)
    post:
      consumes:
      - application/json
      description: Create user Admin or simple
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: create user data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/system.CreateUserData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Create user
      tags:
      - System(Users)
  /system/users/{uid}:
    delete:
      consumes:
      - application/json
      description: Delete simple user
      parameters:
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Delete simple user
      tags:
      - System(Users)
  /system/users/{uid}/password:
    post:
      consumes:
      - application/json
      description: Change user password by admin
      parameters:
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      - description: user new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/system.ChangeUserPassword'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.UsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: Change user password by admin
      tags:
      - System(Users)
  /system/users/login:
    post:
      consumes:
      - application/json
      description: Admin users login with Google captcha(captcha site key required
        in get config api)
      parameters:
      - description: login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/system.LoginData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.UserLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Admin users login
      tags:
      - System(Users)
  /system/users/lookup:
    get:
      consumes:
      - application/json
      description: List of Users Lookup
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UsersLookup'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middlewares.PermissionDenied'
      summary: List of Users Lookup
      tags:
      - System(Users)
  /system/users/password:
    post:
      consumes:
      - application/json
      description: Change user password by self
      parameters:
      - description: user new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/system.ChangeUserPasswordBySelf'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/system.UsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Change user password by self
      tags:
      - System(Users)
  /system/users/profile:
    get:
      consumes:
      - application/json
      description: Get User Profile
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Get User Profile
      tags:
      - System(Users)
swagger: "2.0"